+++ demo/main.qml
@@ -25,14 +25,20 @@ ApplicationWindow {
             "Bottom Sheet", "Dialog", "Forms", "List Items", "Page Stack"
     ]
 
+    property var sections: [ styles, basicComponents, compoundComponents ]
+
+    property var sectionTitles: [ "Style", "Basic Components", "Compound Components" ]
+
+    property string selectedComponent: styles[0]
+
     initialPage: Page {
         id: page
 
         title: "Material Demo"
 
-        tabs: [ "Style", "Basic Components", "Compound Components" ]
+        tabs: navDrawer.enabled ? [] : sectionTitles
 
-        actionBar.maxActionCount: 4
+        actionBar.maxActionCount: navDrawer.enabled ? 3 : 4
 
         actions: [
             Action {
@@ -71,22 +77,66 @@ ApplicationWindow {
             }
         ]
 
+        backAction: navDrawer.action
+
+        NavigationDrawer {
+            id: navDrawer
+
+            enabled: page.width < Units.dp(500)
+
+            Flickable {
+                anchors.fill: parent
+
+                contentHeight: Math.max(content.implicitHeight, height)
+
+                Column {
+                    id: content
+                    anchors.fill: parent
+
+                    Repeater {
+                        model: sections
+
+                        delegate: Column {
+                            width: parent.width
+
+                            ListItem.Subheader {
+                                text: sectionTitles[index]
+                            }
+
+                            Repeater {
+                                model: modelData
+                                delegate: ListItem.Standard {
+                                    text: modelData
+                                    selected: modelData == demo.selectedComponent
+                                    onClicked: {
+                                        demo.selectedComponent = modelData
+                                        navDrawer.close()
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
         TabView {
             id: tabView
             anchors.fill: parent
             currentIndex: page.selectedTab
-            model: [
-                styles, basicComponents, compoundComponents
-            ]
+            model: sections
 
             delegate: Item {
                 width: tabView.width
                 height: tabView.height
+                clip: true
 
                 property string selectedComponent: modelData[0]
 
                 Sidebar {
                     id: sidebar
+
+                    expanded: !navDrawer.enabled
 
                     Column {
                         width: parent.width
@@ -114,8 +164,15 @@ ApplicationWindow {
                     Loader {
                         id: example
                         anchors.fill: parent
+
                         // selectedComponent will always be valid, as it defaults to the first component
-                        source: Qt.resolvedUrl("%1Demo.qml").arg(selectedComponent.replace(" ", ""))
+                        source: {
+                            if (navDrawer.enabled) {
+                                return Qt.resolvedUrl("%1Demo.qml").arg(demo.selectedComponent.replace(" ", ""))
+                            } else {
+                                return Qt.resolvedUrl("%1Demo.qml").arg(selectedComponent.replace(" ", ""))
+                            }
+                        }
                     }
                 }
                 Scrollbar {